Initial irq stuff

+++ src/irq.h
@@ -8,6 +8,6 @@ extern int IRQ_get_mask(int IRQline);
 extern void IRQ_end_of_interrupt(int irq);
 
 typedef void (*irq_handler_t)(int, int, void *);
-extern void IRQ_set_handler(int irq, irq_handler_t handler, void *arg);
+extern void IRQ_set_handler(uint8_t irq, irq_handler_t handler, void *arg);
 
 #endif
